requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
mysqlclient==2.1.1
Werkzeug==2.2.2